LayoutTests/platform/ios-simulator/editing/selection/select-bidi-run-expected.txt

This test ensures WebKit lets user select bidirectional text intuitively. To manually test, select text in blue box in each test case below by a mouse drag from left to right. The changes in the selected text should match the expectations before |. Do the same by a mouse drag from right to left and expectations are after |.

Test "abcABC" in "abcABC":
Selecting from left to right
Selecting from right to left

Test "ABCdef" in "ABCdef":
Selecting from left to right
Selecting from right to left

Test "ABC" in "abcABCdef":
Selecting from left to right
Selecting from right to left

Test "ABC" in "ABCdef":
Selecting from left to right
Selecting from right to left

Test "ef" in "ABCdef":
Selecting from left to right
Selecting from right to left

Test "AB" in "abcABC":
Selecting from left to right
Selecting from right to left

Test "12" in "aXM12JNd":
Selecting from left to right
Selecting from right to left

Test "ABC 123" in "ABC 123":
Selecting from left to right
Selecting from right to left

Test "ABC 123" in "ABC 123":
Selecting from left to right
Selecting from right to left

Test "ABC" in "aABCb":
Selecting from left to right
Selecting from right to left

Test "ABC" in "ABC":
Selecting from left to right
Selecting from right to left

PASS successfullyParsed is true

TEST COMPLETE
